scsi: qla2xxx: Initialize number of active target queues
num_act_qpairs is zero despite a number of queue pairs available to
target. There's an unused constant that can be re-used as default.

The constant is fine because an increase of target queue pairs to 2
gives almost x2 boost of IOPS on nullblk (if target executes commands
simultaneously on both queue pairs).

Signed-off-by: Roman Bolshakov <r.bolshakov@yadro.com>