target: make completion affinity configurable
It may not always be best to complete the IO on same CPU as it was
submitted on. This allows userspace to config it.

This has been useful for vhost-scsi where we have a single thread for
submissions and completions. If we force the completion on the
submission cpu we may be adding conflicts with what the user has setup
in the lower levels with settings like the block layer rq_affinity or
the driver's irq or softirq (the network's rps_cpus value) settings.

We may also want to set it up where the vhost thread runs on CPU N
and does it's submissions/completions there, and then have LIO do
it's completion booking on CPU M, but can't config the lower levels
due to issues like using dm-multipath with lots of paths (the path
selector can throw commands all over the system because it's only
aking into account latency/throughput at its level).

The new setting is in
/sys/kernel/config/target/$fabric/$target/param/cmd_completion_affinity

Writing:
-1  -> gives the current default behavior of completing on the submission
CPU.
-2  -> completes the cmd on the CPU the lower layers sent it to us from.
0 > -> complete on the CPU userspace has specified.

Signed-off-by: Mike Christie <michael.christie@oracle.com>
Reviewed-by: Himanshu Madhani <himanshu.madhani@oracle.com>