scsi: qla2xxx: Use In-Pointer for response queue
The existing way of response queue processing relies on the assumption
that 64-bit DMA is performed atomically for queue elements, i.e. there
could not be an IOCB in the queue that is partially modified by DMA
engine of the HBA - when some IOCB fields are new and others are from
the previous IOCBs.

The assumption doesn't seem to be valid. It was observed that
occasionally qla2xxx can't find command handles in target mode:

  qla_target(0): Suspicious: unable to find the command with handle
  2002056e req->id 2 rsp->id 2

A command in the session would be then forever stuck in cmd_sent_to_fw
state and won't be ever completed because CTIO completion is missed. ACL
deletion, link reinit or OS reboot wouldn't succeed because the stuck
command blocks session deletion.

An attempt to dump IOCB of the "Suspicious" command showed that it
actually contains valid handle. I.e. there's a race and invalid handle
was captured in qlt_response_pkt() from the partially stale IOCB but DMA
transfer is completed by the the moment of IOCB dump.

The handle is passed from register to register after it's read from
entry->handle in qlt_response_pkt() therefore memory corruptions were
ruled out as a possible cause of the behaviour.

The correct way to avoid incomplete IOCBs is to use Response Queue
In-Pointer rather than checking if the tail of IOCB was updated.

Suggested-by: Dmitriy Bogdanov <d.bogdanov@yadro.com>
Signed-off-by: Roman Bolshakov <r.bolshakov@yadro.com>